requirements.txt
streamlit
requests
datetime
google-generativeai
plotly
